+++ src/main/java/fuzs/bettermodsbutton/BetterModsButton.java
@@ -1,12 +1,16 @@
 package fuzs.bettermodsbutton;
 
 import fuzs.bettermodsbutton.config.ClientConfig;
-import fuzs.bettermodsbutton.config.core.AbstractConfig;
-import fuzs.bettermodsbutton.config.core.ConfigHolder;
-import fuzs.bettermodsbutton.config.core.ConfigHolderImpl;
+import fuzs.bettermodsbutton.lib.config.AbstractConfig;
+import fuzs.bettermodsbutton.lib.config.ConfigHolder;
+import fuzs.bettermodsbutton.lib.config.ConfigHolderImpl;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.ExtensionPoint;
+import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
+import net.minecraftforge.fml.network.FMLNetworkConstants;
+import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -24,5 +28,6 @@ public class BetterModsButton {
     @SubscribeEvent
     public static void onConstructMod(final FMLConstructModEvent evt) {
         ((ConfigHolderImpl<?, ?>) CONFIG).addConfigs(MODID);
+        ModLoadingContext.get().registerExtensionPoint(ExtensionPoint.DISPLAYTEST, () -> Pair.of(() -> FMLNetworkConstants.IGNORESERVERONLY, (remote, isServer) -> true));
     }
 }