+++ src/main/java/fuzs/bettermodsbutton/config/ClientConfig.java
@@ -1,8 +1,8 @@
 package fuzs.bettermodsbutton.config;
 
 import fuzs.bettermodsbutton.client.handler.ModScreenHandler;
-import fuzs.bettermodsbutton.config.core.AbstractConfig;
-import fuzs.bettermodsbutton.config.core.annotation.Config;
+import fuzs.bettermodsbutton.lib.config.AbstractConfig;
+import fuzs.bettermodsbutton.lib.config.annotation.Config;
 
 public class ClientConfig extends AbstractConfig {
     @Config(name = "main_menu_mods_button", description = "Where to place mods button on main menu screen.")
@@ -13,6 +13,8 @@ public class ClientConfig extends AbstractConfig {
     public ModScreenHandler.PauseScreenMode pauseScreenMode = ModScreenHandler.PauseScreenMode.INSERT_BELOW_FEEDBACK_AND_BUGS;
     @Config(description = "Show a small green orb when mod updates are available.")
     public boolean updateNotification = false;
+    @Config(description = "Force a dirt background to show for the mod list screen, even when inside of a world where it would normally be transparent.")
+    public boolean forceDirtBackground = true;
 
     public ClientConfig() {
         super("");