+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v1.0.3-1.16.5] - 2021-11-02
+### Changed
+- Added dirt background for mod list when opened from inside a world
+- Servers no longer show an incompatible mods list
+- Compatibility with [Catalogue] mod
+
 ## [v1.0.2-1.16.5] - 2021-10-29
 ### Added
 - Support for inner configs for annotation system
@@ -16,3 +22,4 @@ The format is based on [Keep a Changelog].
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Puzzles Lib]: https://www.curseforge.com/minecraft/mc-mods/puzzles-lib
+[Catalogue]: https://www.curseforge.com/minecraft/mc-mods/catalogue